Integrate proxy command into tunnel
